Fix delete issue caused by []byte marshall
Based on suggested fix in #104

Delete's would not work properly because of the way the params were
passed through. If an empty list of params was passed, it would actually
Marshall that to a byte string like [34 34]

This is representative of '""' in the body of a curl http call. The
delete function requires no params so this may really be oboslete, but
this is a good workaround for now.