client/CreateQemuVM: only read body if existing
Working on Packer, we noticed that on many occasions the plugin crashes
because the call to `resp.Body.Close()` panics with a null pointer
dereference.

Unsure if this is because the `resp` object is null, or if the
`resp.Body` is, but to avoid this problem, we only read and defer
closing the body when it is available.